web/satellite/jest.config: Always use UTC
Update jest config to use UTC so that undesired timestamp changes are
not made to snapshots when they are updated in American timezones.

Right now, if you are in an American timezone, there are diffs like this
in the snapshot files after updating them:

-  <p class="user-container__date">1/1/1970</p>
+  <p class="user-container__date">12/31/1969</p>

The hacky solution is to manually set timezone to Europe/Amsterdam in
the terminal before updating snapshots, but this commit should fix that
automatically.

Change-Id: I7cbbda0493b229fcd22b863dd688201436edc18a